app/requirements.txt
# Core Framework
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.11.0
python-dotenv==1.0.0

# LangGraph & Agents
langgraph==0.2.45
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.29
crewai==0.201.1
crewai-tools==0.75.0

# Memory & Vectors
weaviate-client==4.9.3
qdrant-client==1.11.1

# LLM Providers
ollama==0.5.3
openai==1.99.9
anthropic==0.66.0
litellm==1.76.2

# File Processing
Pygments==2.17.2
pdf2image==1.17.0
Pillow>=10.0.0
python-magic-bin==0.4.14
PyPDF2==3.0.1

# Utilities
httpx==0.28.1
pydantic==2.11.10
numpy==1.26.4
aiohttp==3.9.1

# Testing
pytest==8.3.4
pytest-asyncio==0.23.5
